Vector Job: combinact9
Applies combinact activation with shuffled permutations. Shuffle maps
are decided when model is initialized: for each hidden layer, we
specify a unique shuffle map for every one if its permutations. A single
shuffle map defines a random permutation of the indices of the original
inputs to the layer. This way every permutation in every layer provides
a unique random shuffling.

While every permutation provides a unique shuffling, the shufflings
remain constant across mini-batches and epochs (ie. the network
will always shuffle the first permutation of the first layer in the
same way). This way the network knows what shuffles to expect and can
learn relationships between the shuffled parameters.